founders-fund-app/ocr-worker/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
Pillow
pytesseract
google-cloud-vision
